Add uiscene and message files
