+++ src/Menus/HeroesMenu.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-useless-constructor */
 import Menu from './Menu';
 
 export default class HeroesMenu extends Menu {
@@ -5,3 +6,4 @@ export default class HeroesMenu extends Menu {
     super(x, y, scene);
   }
 }
+/* eslint-enable no-useless-constructor */
