+++ src/Scenes/BattleScene.js
@@ -1,11 +1,8 @@
-/* eslint-disable no-undef, no-plusplus, max-classes-per-file */
+/* eslint-disable no-undef, no-plusplus */
 import PlayerCharacter from '../Units/Player';
 import Enemy from '../Units/Enemy';
-import EnemiesMenu from '../Menus/EnemiesMenu';
-import ActionsMenu from '../Menus/ActionsMenu';
-import HeroesMenu from '../Menus/HeroesMenu';
 
-class BattleScene extends Phaser.Scene {
+export default class BattleScene extends Phaser.Scene {
   constructor() {
     super('BattleScene');
   }
@@ -120,151 +117,4 @@ class BattleScene extends Phaser.Scene {
     // return to WorldScene and sleep current BattleScene
   }
 }
-
-// User Interface scene
-class UIScene extends Phaser.Scene {
-  constructor() {
-    super('UIScene');
-  }
-
-  create() {
-    this.graphics = this.add.graphics();
-    this.graphics.lineStyle(1, 0xffffff);
-    this.graphics.fillStyle(0x031f4c, 1);
-
-    // rectangle 1
-    this.graphics.strokeRect(2, 250, 140, 100);
-    this.graphics.fillRect(2, 250, 140, 100);
-    // rectangle 2
-    this.graphics.strokeRect(145, 250, 130, 100);
-    this.graphics.fillRect(145, 250, 130, 100);
-    // rectangle 3
-    this.graphics.strokeRect(278, 250, 140, 100);
-    this.graphics.fillRect(278, 250, 140, 100);
-
-    // menus
-    this.menus = this.add.container();
-
-    this.heroesMenu = new HeroesMenu(295, 263, this);
-    this.actionsMenu = new ActionsMenu(178, 283, this);
-    this.enemiesMenu = new EnemiesMenu(18, 263, this);
-
-    this.currentMenu = this.actionsMenu;
-
-    this.menus.add(this.heroesMenu);
-    this.menus.add(this.actionsMenu);
-    this.menus.add(this.enemiesMenu);
-
-    // getting heroes and enemies from battleScene
-    this.battleScene = this.scene.get('BattleScene');
-
-    // cursors
-    this.input.keyboard.on('keydown', this.onKeyInput, this);
-
-    // listener for turn-events
-    this.battleScene.events.on('PlayerSelect', this.onPlayerSelect, this);
-    this.events.on('SelectedAction', this.onSelectedAction, this);
-    this.events.on('Enemy', this.onEnemy, this);
-
-    // wake event listener
-    this.sys.events.on('wake', this.createMenu, this);
-
-    // messages
-    this.message = new Message(this, this.battleScene.events);
-    this.add.existing(this.message);
-
-    this.createMenu();
-  }
-
-  createMenu() {
-    this.remapHeroes();
-    this.remapEnemies();
-
-    this.battleScene.nextTurn();
-  }
-
-  remapHeroes() {
-    const { heroes } = this.battleScene;
-    this.heroesMenu.remap(heroes);
-  }
-
-  remapEnemies() {
-    const { enemies } = this.battleScene;
-    this.enemiesMenu.remap(enemies);
-  }
-
-  onKeyInput(event) {
-    if (this.currentMenu && this.currentMenu.selected) {
-      if (event.code === 'ArrowUp') {
-        this.currentMenu.moveSelectionUp();
-      } else if (event.code === 'ArrowDown') {
-        this.currentMenu.moveSelectionDown();
-      } else if (event.code === 'Space' || event.code === 'ArrowLeft') {
-        this.currentMenu.confirm();
-      }
-    }
-  }
-
-  onPlayerSelect(id) {
-    this.heroesMenu.select(id);
-    this.actionsMenu.select(0);
-    this.currentMenu = this.actionsMenu;
-  }
-
-  onSelectedAction() {
-    this.currentMenu = this.enemiesMenu;
-    this.enemiesMenu.select(0);
-  }
-
-  onEnemy(index) {
-    this.heroesMenu.deselect();
-    this.actionsMenu.deselect();
-    this.enemiesMenu.deselect();
-    this.currentMenu = null;
-    this.battleScene.receivePlayerSelection('attack', index);
-  }
-}
-
-// the message class extends containter
-class Message extends Phaser.GameObjects.Container {
-  constructor(scene, events) {
-    super(scene, 160, 30);
-    const graphics = this.scene.add.graphics();
-    this.add(graphics);
-
-    graphics.lineStyle(1, 0xffffff, 0.8);
-    graphics.fillStyle(0x031f4c, 0.3);
-    graphics.strokeRect(-50, -15, 200, 30);
-    graphics.fillRect(-50, -15, 200, 30);
-
-    this.text = new Phaser.GameObjects.Text(scene, 50, 0, '', {
-      color: '#ffffff',
-      align: 'center',
-      fontSize: 12,
-      wordWrap: { width: 195, useAdvancedWrap: true },
-    }).setOrigin(0.5);
-    this.add(this.text);
-
-    events.on('Message', this.showMessage, this);
-    this.visible = false;
-  }
-
-  showMessage(text) {
-    this.text.setText(text);
-    this.visible = true;
-    if (this.hideEvent) {
-      this.hideEvent.remove(false);
-    }
-    this.hideEvent = this.scene.time.addEvent(
-      { delay: 2000, callback: this.hideMessage, callbackScope: this },
-    );
-  }
-
-  hideMessage() {
-    this.hideEvent = null;
-    this.visible = false;
-  }
-}
-
-export { BattleScene, UIScene };
 /* eslint-enable no-undef, no-plusplus */
